+++ nadeko_master_installer.sh
@@ -12,6 +12,7 @@
 #          auto restart.
 #   E.1. - Return to prevent further code execution.
 #   F.1. - Set the execution permissions for the downloaded script, then execute it.
+#   G.1. - Prevent the code from running if the options is disabled.
 #
 ########################################################################################
 #### [ Variables and Functions ]
@@ -279,7 +280,6 @@ else
         #           service)
         # @param $2 Dictates whether or not text indicating that the service has been
         #           stopped or is currently stopped, should be printed to the terminal.
-        #
         ####
 
         case "$1" in
@@ -506,15 +506,16 @@ fi
 echo -e "Welcome to the NadekoBot installer\n"
 
 while true; do
-    ## Create '$nadeko_service_name', if it does not already exist.
+    ## Create $nadeko_service_name, if it does not already exist.
     if [[ ! -f $nadeko_service ]]; then
         echo "Creating '$nadeko_service_name'..."
-        ## Create '/Users/"$USER"/Library/LaunchAgents' if 'LaunchAgents' doesn't
-        ## already exist.
+        ## If running on macOS, create '/Users/"$USER"/Library/LaunchAgents' if
+        ## 'LaunchAgents' doesn't already exist.
         if [[ $_DISTRO = "Darwin" && ! -d /Users/$USER/Library/LaunchAgents/ ]]; then
             # TODO: Add error catching???
             mkdir /Users/"$USER"/Library/LaunchAgents
         fi
+        # TODO: Add comments to this echo...
         echo -e "$nadeko_service_content" | sudo tee "$nadeko_service" &>/dev/null &&
             if [[ $_DISTRO != "Darwin" ]]; then
                 # Make changes to services in 
@@ -530,52 +531,60 @@ while true; do
             }
     fi
 
+    # Get the current status of $nadeko_service_name.
     service_actions "nadeko_service_status"
 
     ################################################################################
     ##### [ User Options For Starting NadekoBot ]
 
-
-    if (! hash dotnet || ! hash redis-server || ! hash git || ! \
-            hash jq || (! hash python && ! hash python3) || ! hash \
-            youtube-dl) &>/dev/null; then
-        disabled_1=true
+    ## Disable option 1, if any of the following tools are not installed.
+    if (! hash dotnet || ! hash redis-server || ! hash git || ! hash jq || (! hash \
+            python && ! hash python3) || ! hash youtube-dl) &>/dev/null; then
+        option_one_disabled=true
         echo "${_GREY}1. Download NadekoBot (Disabled until option 6 is ran)$_NC"
+    ## Else enable it.
     else
-        disabled_1=false
+        option_one_disabled=false
         echo "1. Download NadekoBot"
     fi
 
+    ## Disable options 2, 3, and 5 if any of the tools are not installed or non of the
+    ## directories/files could be found.
     if [[ ! -d NadekoBot/src/NadekoBot/ || ! -f NadekoBot/src/NadekoBot/credentials.json ||
-            ! -d NadekoBot/src/NadekoBot/bin/Release || -z $(jq -r ".Token" NadekoBot/src/NadekoBot/credentials.json) ]] ||
-            (! hash dotnet || ! hash redis-server || ! hash git || ! \
-            hash jq || (! hash python && ! hash python3) || ! hash \
-            youtube-dl) &>/dev/null; then
-        disabled_23=true
-        disabled_5=true
-        echo "${_GREY}2. Run NadekoBot in the background (Disabled" \
-            "until options 1, 6, and 6 are ran)"
-        echo "3. Run NadekoBot in the background with auto" \
-            "restart (Disabled until options 1, 6, and 6 are ran)$_NC"
-    elif [[ -f NadekoRun.sh ]]; then
-        disabled_23=false
+            ! -d NadekoBot/src/NadekoBot/bin/Release || 
+            -z $(jq -r ".Token" NadekoBot/src/NadekoBot/credentials.json) ]] ||
+            (! hash dotnet || ! hash redis-server || ! hash git || ! hash jq || 
+            (! hash python && ! hash python3) || ! hash  youtube-dl) &>/dev/null; then
+        option_two_and_three_disabled=true
+        option_five_disabled=true
 
+        echo "${_GREY}2. Run NadekoBot in the background (Disabled until options 1," \
+            "6, and 6 are ran)"
+        echo "3. Run NadekoBot in the background with auto restart (Disabled until" \
+            "options 1, 6, and 6 are ran)$_NC"
+    ## Enable the ability to run NadekoBot in any run mode, if 'NadekoRun.sh' does not
+    ## exist.
+    elif [[ -f NadekoRun.sh ]]; then
+        option_two_and_three_disabled=false
+
+        ## Enable option 5 if NadekoBot's service is running.
         if [[ $nadeko_service_status = "active" || $nadeko_service_status = "running" ]]; then
-            disabled_5=false
+            option_five_disabled=false
             run_mode_status=" ${_GREEN}(Running in this mode)$_NC"
+        ## Disable option 5 if NadekoBot's service not running.
         elif [[ $nadeko_service_status = "inactive" || $nadeko_service_status = "waiting" ]]; then
-            disabled_5=true
+            option_five_disabled=true
             run_mode_status=" ${_YELLOW}(Set up to run in this mode)$_NC"
+        ## Else don't do anything...
         else
             run_mode_status=" ${_YELLOW}(Status unkown)$_NC"
         fi
 
-        if [[ $(grep '_code_name_="NadekoRunARU"' NadekoRun.sh) ]]; then
-            echo "2. Run NadekoBot in the background"
-            echo "3. Run NadekoBot in the background with auto restart"
-        elif [[ $(grep '_code_name_="NadekoRunAR"' NadekoRun.sh) ]]; then
+        ## If NadekoBot is running in the background with auto restart...
+        if [[ $(grep '_code_name_="NadekoRunAR"' NadekoRun.sh) ]]; then
             echo "2. Run NadekoBot in the background"
             echo "3. Run NadekoBot in the background with auto restart${run_mode_status}"
+        ## Else if NadekoBot is running in the background...
         elif [[ $(grep '_code_name_="NadekoRun"' NadekoRun.sh) ]]; then
             echo "2. Run NadekoBot in the background${run_mode_status}"
             echo "3. Run NadekoBot in the background with auto restart"
@@ -583,16 +592,17 @@ while true; do
             echo "2. Run NadekoBot in the background"
             echo "3. Run NadekoBot in the background with auto restart"
         fi
+    ## Enable options 2 and 3, but disable option 5.
     else
-        disabled_23=false
-        disabled_5=true
+        option_two_and_three_disabled=false
+        option_five_disabled=true
         echo "2. Run NadekoBot in the background"
         echo "3. Run NadekoBot in the background with auto restart"
     fi
 
     echo "4. Stop NadekoBot"
 
-    if [[ $disabled_5 = true ]]; then
+    if [[ $option_five_disabled = true ]]; then
         echo "${_GREY}5. Display '$nadeko_service_name' logs in follow mode" \
             "(Disabled until NadekoBot has been started)$_NC"
     else
@@ -601,19 +611,21 @@ while true; do
 
     echo "6. Install prerequisites"
 
+    ## Disable option 7 if NadekoBot has not been downloaded.
     if [[ ! -d NadekoBot/src/NadekoBot/ ]]; then
         echo "${_GREY}7. Set up credentials.json (Disabled until option 1 is ran)$_NC"
-        disabled_7=true
+        option_seven_disabled=true
     else
         echo "7. Set up credentials.json"
-        disabled_7=false
+        option_seven_disabled=false
     fi
 
     echo "8. Exit"
     read -r choice
     case "$choice" in
-        1)
-            if [[ $disabled_1 = true ]]; then
+        1)  
+            ## G.1. 
+            if [[ $option_one_disabled = true ]]; then
                 clear -x
                 echo "${_RED}Option 1 is currently disabled$_NC"
                 continue
@@ -626,13 +638,13 @@ while true; do
             if [[ $_DISTRO != "Darwin" ]]; then export nadeko_service_content; fi
 
             echo "Downloading 'nadeko_latest_installer.sh'..."
-            # Download latest version of 'nadeko_latest_installer.sh'.
+            # Download the latest version of 'nadeko_latest_installer.sh'.
             curl -s "$_RAW_URL"/nadeko_latest_installer.sh -o nadeko_latest_installer.sh || {
                 echo "${_RED}Failed to download latest 'nadeko_latest_installer.sh'...$_NC" >&2
                 _CLEAN_EXIT "1" "Exiting" "true"
             }
-            # B.1. Clears screen of current content.
             clear -x
+            # F.1.
             sudo chmod +x nadeko_latest_installer.sh && ./nadeko_latest_installer.sh
 
             # Rexecutes the new/downloaded version of 'installer_prep.sh', so that all
@@ -641,7 +653,8 @@ while true; do
             ;;
         2)
             clear -x
-            if [[ $disabled_23 = true ]]; then
+            ## G.1. 
+            if [[ $option_two_and_three_disabled = true ]]; then
                 echo "${_RED}Option 2 is currently disabled$_NC"
                 continue
             fi
@@ -653,7 +666,8 @@ while true; do
             ;;
         3)
             clear -x
-            if [[ $disabled_23 = true ]]; then
+            ## G.1. 
+            if [[ $option_two_and_three_disabled = true ]]; then
                 echo "${_RED}Option 3 is currently disabled$_NC"
                 continue
             fi
@@ -673,7 +687,8 @@ while true; do
             ;;
         5)
             clear -x
-            if [[ $disabled_5 = true ]]; then
+            ## G.1. 
+            if [[ $option_five_disabled = true ]]; then
                 echo "${_RED}Option 5 is currently disabled$_NC"
                 continue
             fi
@@ -687,8 +702,8 @@ while true; do
                     "+ C'\n2) Press 'Q'$_NC"
             fi
             if [[ $_DISTRO != "Darwin" ]]; then
-                # The pipe makes it possible to exit journalctl without exiting
-                # the script
+                # The pipe makes it possible to exit journalctl without exiting the
+                # script
                 sudo journalctl -f -u "$nadeko_service_name"  | less -FRSXM
             else
                 tail -f "bot.nadeko.Nadeko.log" | less -FRSXM
@@ -697,17 +712,19 @@ while true; do
             ;;
         6)
             echo "Downloading '$prereqs_installer' as 'prereqs_installer.sh'..."
-            # Download latest version of '$prereqs_installer'.
+            # Download latest version of $prereqs_installer.
             curl -s "$_RAW_URL"/"$prereqs_installer" -o prereqs_installer.sh || {
                 echo "${_RED}Failed to download latest '$prereqs_installer'...$_NC" >&2
                 _CLEAN_EXIT "1" "Exiting" "true"
             }
-            clear -x  # B.1.
+            clear -x
+            # F.1.
             sudo chmod +x prereqs_installer.sh && ./prereqs_installer.sh
-            clear -x  # B.1.
+            clear -x
             ;;
         7)
-            if [[ $disabled_7 = true ]]; then
+            ## G.1. 
+            if [[ $option_seven_disabled = true ]]; then
                 clear -x
                 echo "${_RED}Option 7 is currently disabled$_NC"
                 continue
@@ -719,9 +736,10 @@ while true; do
                 echo "${_RED}Failed to download latest 'credentials_setup.sh'...$_NC" >&2
                 _CLEAN_EXIT "1" "Exiting" "true"
             }
-            clear -x  # B.1.
+            clear -x
+            # F.1.
             sudo chmod +x credentials_setup.sh && ./credentials_setup.sh
-            clear -x  # B.1.
+            clear -x
             ;;
         8)
             _CLEAN_EXIT "0" "Exiting"