+++ CHANGELOG.md
@@ -6,14 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-**IMPORTANT**
+### Summary (IMPORTANT)
+
 Due to some breaking changes inside of 'linuxAIO.sh', all users who are currently using 'linuxAIO.sh' revision 8 and earlier will receive a message the next time they execute the script. The message will inform users that they will need to manually download the newest version of 'linuxAIO.sh'. The appropriate command to do this will be provided by the installer, based on the current configurations of 'linuxAIO.sh'.
 
 Additionally, you'll need to delete `/lib/systemd/system/nadeko.service`, as the service will now be stored in `/etc/systemd/system/nadeko.service`. To do this, run the following command: `sudo systemctl stop nadeko.service && sudo rm /lib/systemd/system/nadeko.service && sudo systemctl daemon-reload`. From here, start NadekoBot like you have any other time.
 
 ### Added
 
-- End user can now configure what branch/tag to have the installer use when downloading NadekoBot.
+- End user can now configure what branch/tag to download NadekoBot from (i.e. `1.9`, `2.39.1`, etc.).
 - Now sets `$allow_run_as_root` and `$_NADEKO_INSTALL_VERSION` the same way that `$installer_branch` is set, whenever a new version of 'linuxAIO.sh' is retrieved.
 
 ### Changed